+++ app/src/main/java/dev/kostromdan/mods/crash_assistant/app/gui/FilePanel.java
@@ -283,7 +283,7 @@ public class FilePanel {
 
     public String getMessageWithBothLinks(boolean forMsg) {
         Function<String, String> langFunc = forMsg ? LanguageProvider::getMsgLang : LanguageProvider::get;
-        return getFileName() + ": " + "[`" + langFunc.apply("gui.split_log_dialog_head").toLowerCase() + "`](<" + getUploadedLinkFirstLines() + ">) / " +
+        return "[`" + getFileName() + " " + langFunc.apply("gui.split_log_dialog_head").toLowerCase() + "`](<" + getUploadedLinkFirstLines() + ">) / " +
                 "[`" + langFunc.apply("gui.split_log_dialog_tail").toLowerCase() + "`](<" + getUploadedLinkLastLines() + ">) " + getTooBigReasons(forMsg) + "\n";
     }