+++ common_config/src/main/java/dev/kostromdan/mods/crash_assistant/config/CrashAssistantConfig.java
@@ -113,10 +113,6 @@ public class CrashAssistantConfig {
                         "Supports html formatting, placeholders.\n" +
                         "Leave empty to prevent showing this warning message.",
                 "");
-        addOption("generated_message.generated_msg_includes_info_why_split",
-                "Crash Assistant splits the log into a head and tail of 25,000 lines each if the log exceeds mclo.gs limits (10 MB or 25,000 lines).\n" +
-                        "Add an explanation at the bottom of the generated message about why log was split.",
-                true);
 
         config.setComment("modpack_modlist", "Settings of modlist feature.\n" +
                 "Adds in generated msg block about which mods modpack user added/removed/updated.\n" +